+++ demosplan/DemosPlanCoreBundle/MessageHandler/PurgeDeletedProceduresMessageHandler.php
@@ -11,8 +11,10 @@
 namespace demosplan\DemosPlanCoreBundle\MessageHandler;
 
 use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
+use DemosEurope\DemosplanAddon\Contracts\PermissionsInterface;
 use demosplan\DemosPlanCoreBundle\Logic\Procedure\ProcedureHandler;
 use demosplan\DemosPlanCoreBundle\Message\PurgeDeletedProceduresMessage;
+use demosplan\DemosPlanCoreBundle\Traits\InitializesAnonymousUserPermissionsTrait;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
@@ -20,15 +22,20 @@ use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 #[AsMessageHandler]
 final class PurgeDeletedProceduresMessageHandler
 {
+    use InitializesAnonymousUserPermissionsTrait;
+
     public function __construct(
         private readonly ProcedureHandler $procedureHandler,
         private readonly GlobalConfigInterface $globalConfig,
+        private readonly PermissionsInterface $permissions,
         private readonly LoggerInterface $logger,
     ) {
     }
 
     public function __invoke(PurgeDeletedProceduresMessage $message): void
     {
+        $this->initializeAnonymousUserPermissions();
+
         $this->logger->info('Purge deleted procedures... ');
         $purgedProcedures = 0;
         try {