+++ demosplan/DemosPlanCoreBundle/MessageHandler/SwitchElementStatesMessageHandler.php
@@ -10,8 +10,10 @@
 
 namespace demosplan\DemosPlanCoreBundle\MessageHandler;
 
+use DemosEurope\DemosplanAddon\Contracts\PermissionsInterface;
 use demosplan\DemosPlanCoreBundle\Logic\Document\ElementsService;
 use demosplan\DemosPlanCoreBundle\Message\SwitchElementStatesMessage;
+use demosplan\DemosPlanCoreBundle\Traits\InitializesAnonymousUserPermissionsTrait;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
@@ -19,14 +21,19 @@ use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 #[AsMessageHandler]
 final class SwitchElementStatesMessageHandler
 {
+    use InitializesAnonymousUserPermissionsTrait;
+
     public function __construct(
         private readonly ElementsService $elementService,
+        private readonly PermissionsInterface $permissions,
         private readonly LoggerInterface $logger,
     ) {
     }
 
     public function __invoke(SwitchElementStatesMessage $message): void
     {
+        $this->initializeAnonymousUserPermissions();
+
         $this->logger->info('switchStatesOfToday');
 
         $affectedElements = 0;