+++ client/js/components/procedure/admin/AdministrationCustomFieldsList.vue
@@ -3,6 +3,7 @@
     <dp-inline-notification
       v-if="isStatementField && statementsCount > 0"
       class="mb-4"
+      data-cy="customFields:editInfo"
       dismissible
       :dismissible-key="helpTextDismissibleKey"
       type="warning"